requirements.txt
pillow==11.0.0
tqdm==4.67.0
ftfy==6.3.1
regex==2024.11.6
pandas==2.2.3 
scikit-learn==1.5.2  
timm==1.0.11
matplotlib==3.9.2
torchmetrics==1.5.2